Use the SearchText from the extension (#319)
And the placeholder text too.

This one's mildly confusing, because we don't want the View and the Model fighting over the same exact property. Basically I just have the view be able to call the SearchText setter, and the model's PropChanged is listened to by the View. They shouldn't fight too much then. 

Closes #163
Closes #112