Add support for hiding messages too (#355)
Bits of status messages that were omitted from #281. 

This lets extensions hide messages (and exposes the helper in the helper lib). 

It also adds support for displaying progress as a progress bar underneath the text of the status message. I'll need An Adult to help with the XAML, to re-template the InfoBar to allow a progress wheel in the icon instead, but for now? good enough. 

I'm doing this to unblock the next PR, which should add some rudimentary winget support.