Attempt to correlate apps to winget packages (#387)
This is a first best effort at #367. 

If an app is installed, we'll try to lookup if there's one and only app with that name in our big list of apps. If there is, then we'll set the command on the winget list to the "run" command. 

The uninstall command moves into the context menu, always. Even if we don't find it. This will make it harder to accidentally uninstall things. 

This also has another drive-by fix for #357. Problem wasn't just being on the UI thread, it's that we were trying to get the `.Kind` out of the activation args, after the spawning process had already died.