Implement a basic recently-used-apps algorithm  (#454)
This is a very basic frecency implementation, to let MRU apps and commands show up higher in search results. 

Adds support for a `state.json` file, to store this state. Implication is the same as Terminal:
* `settings.json`: These are _your_ settings, take this with you where you go
* `state.json`: this is state _we_ own. You can take it with you at your peril. **We** control this file. 


This seemed good enough to get "VS" to match to "Visual studio" after just a single launch of VS. 

We can iterate on the weighting in the future. 

Closes #317 
This was also half-tracked in closes #147