Add to the SUI a setting to control where the window is summoned (#461)
Closes #446

Setting is pretty straightforward. Same as PT Run, with an added "leave it where it is", from the Terminal codebase. 

The `DisplayArea` code is net-new (PT used WPF for that, and Terminal used C++), but it's pretty same-y to everything else. 

Drive-by fixes "holy whitespace batman", by making the SUI less wide.