Fixes a race in extension startup
This only repros on my desktop, so I suppose that means a slower machine is needed

I was mistaken, and assumed we were already operating on a copy here. We weren't.
That meant that it was possible for another extension to be detected, change the list,
and crash the whole palette.