Don't auto-hide when an MSAL dialog is opened on us
Apps that want to show MSAL dialogs on the Command Palette would explode
if they passed CmdPal's HWND to WithParentActivityOrWindow. It's not entirely
clear why, but MSAL would explode if the parent HWND is hidden.

When the MSAL dialog opened, we'd hide ourselves, and badda bing, badda boom,
the extension would crash.

MSAL dialogs will set us to WS_DISABLED right before the dialog is opened.
Easy solution. Don't hide ourselves, if we're disabled.

Helps some friends not depend on the existence of Teams :P