Update the settings form when ext settings are saved
This one was subtle - the Settings class in the toolkit didn't ever
change the items for a SettingsContentPage. For the main settings
window, this was problematic. It would only ever hang onto one instance
of that CommandSettings.SettingsContentPage, and never re-retrieve the
value from it.

This fixes that issue, by making sure to raise an ItemsChanged in the
settings changed handler, so that we automatically pull down the new
settings forms.

For settings that were added to commands, as a context item, this wasn't
an issue. They were always returning new forms to the host, with the
current settings values in it.