CmdPal: make actions loading a bit more stable
Mostly stashing this so I remember to come back to it next week.

I was seeing some oddities with actions failing to load on my OS build.
Then I realized that if the actions fail to load, we'll reset the COM
object for the runtime and re-create it. Well if GetActionsForInputs
fails once, it'll probably fail the next time?

anyways, this tries to make that a wee bit more stable.