Extend CommandItemViewModel to handle additional model changes affecting Title
CommandItemViewModel now updates its internal _itemTitle field not only when the CommandItem model raises a property notification for Title, but also when the model’s Command or the Command’s name changes.

This works around an issue in SDK versions before 0.3, where CommandItem does not raise property change notifications for Title in all effective cases.
This approach is not fully correct, as it relies on knowledge of CommandItem’s internal logic and the Title property being virtual and should be sooner or later reverted.