CmdPal: Separate AppHosts from _extension_ app hosts (#40561)
targets: #40560

The other half of the above PR.

`CommandPaletteHost` is a class we use for being the recipient for
status messages, log messages, and other things an extension might want
to communicate. It also very explicitly knows _what extension it's
connected to_. That means it can't live in the `Core` project (because
of that dependency on extensions).

This PR splits that class in half. 
* `AppExtensionHost` does almost all the work from before
* `CommandPaletteHost` is now just a subclass that also knows about
extensions.

Pages now only need an `AppExtensionHost`, not a CPH. 

The wackiest part of the ShellViewModel's `PerformCommand` was the logic
to figure out what the `CommandPaletteHost` for a given page should be.
This moves all that logic out of ShellViewModel, and into a `.UI`
specific class.

ref #40113