Avoid reentrancy issues when loading more items.
When checking the HasMoreItems flag, COM can start a nested message pump, which allows a reentrant call on the XAML UI and causes a fast fail. This change moves the check off the UI thread to prevent reentrancy, but the loading flag is set before we know for sure that there is something to load.

This update also introduces a change: if LoadMore fails, we clear the loading flag immediately.