Handle exceptions when enqueuing callbacks to UI thread in IconCacheService
Handle exceptions thrown in TryEnqueue callbacks so they don’t crash the app (as they cannot be caught by the global exception handler). Any exceptions are now returned to the caller for handling. Additionally, a failure to enqueue the operation onto the dispatcher will also result in an exception.

This is not a breaking change, as exceptions only propagate within the class and do not affect external callers.

Ref: #38260